+++ taguig-suki/app/Http/Requests/UpdateInventoryRequest.php
@@ -14,9 +14,9 @@ class UpdateInventoryRequest extends FormRequest
     public function rules(): array
     {
         return [
-            'stock_quantity' => ['sometimes', 'integer', 'min:0'],
-            'reorder_level' => ['sometimes', 'integer', 'min:0'],
-            'max_stock_level' => ['nullable', 'integer', 'min:0'],
+            'stock_quantity' => ['sometimes', 'numeric', 'min:0'],
+            'reorder_level' => ['sometimes', 'numeric', 'min:0'],
+            'max_stock_level' => ['nullable', 'numeric', 'min:0'],
             'cost_price' => ['nullable', 'numeric', 'min:0'],
             'selling_price' => ['sometimes', 'numeric', 'min:0'],
             'markup_percentage' => ['nullable', 'numeric', 'min:0', 'max:999'],
